Adjusting the interpolation algorithms
Temporal interpolation is not required for the "fine to coarse" mesh interpolation.

Based on the new main algorithm in the previous commit, I split the "coarse to fine" mesh interpolation into two algorithms, one for spatial and one for temporal. The spatial interpolation only needs to be done at the beginning of the sub-iteration process on the fine grid. The temporal interpolation has to be performed at every sub-iteration.