Modifying the Main.f90 algorithm to call the appropriate interpolation routines
Also increased the allocation and computation of feqFF arrays to include 2 nodes onn either side of the fine mesh to account for interpolation requirements.

There seems to be a problem,  nzSub seems to be zero inside AllocateArrays_fine. This is not good. Need to go back a few commits to see where this bug was introduced.