Changed the coarse to fine interpolation to use f instead fPlus
It turns out that the post-collision density distribution is stored only in 'f' first and not in 'fPlus'. At the beginning of the Stream function, we do 'fPlus = f'. Hence while going from the coarse to the fine mesh, we need to use the value stored in 'f' as the post-collision density distribution. However, while going from the fine mesh to the coarse mesh, the streaming operation has already been performed. Hence it's 'fPlus_fine' that contains the true post-collision, pre-streamed density distribution.