Changing the dimensions of the domain in the z direction to start from 0
The dimensions of the coarse mesh originally started at z=0.5 dz. Due to periodicity in the z direction, the number of nodes would correspond to a dimension 1 grid length less than the period in the z direction. Either Yanxing, Gino or Balaji should've decided to split this equally at the beginning and the end of the domain. This is changed such that both coarse and fine meshes now start at z=0 and end such that the next node of the corresponding mesh would correspond to z=0 again due to the periodicity.

In Geometry_fine.f90, the time is calculated based on '(iter-1)+subIter/gridRatio' instead of just 'iter'.