Creating a flag that indicates whether each particle is in the coarse or fine mesh region
This is necessary because the coarse mesh gets the first shot at tracking each particle over a time step and then the fine mesh. If the coarse mesh propagates the particle into the fine mesh region, we do not want the fine mesh to pick it up and track it again. We want the fine mesh to track it only at the next time step. This flag will be incorporated into the Particle_Track routines of both the coarse and the fine mesh.