Using the flag for particle in coarse vs. fine mesh in the Particle_Track subroutines
The flag is set when the coarse mesh goes through the Particle_Track routine. A particle is deemed to be in the coarse mesh if it's completely outside the blue region (hardCheckCoarseMesh) or in the interface region between the two meshes (softCheckCoarseMesh), is coming out of the fine mesh and will hypothetically be in the coarse mesh next time step (checked using first order time-stepping).

The flag also needs to be set when the particle is transferred between two processors. This is usually not an issue if the particle completely remains in the coarse mesh. However, if it is in the fine mesh and gets transferred in between two fine mesh time steps in the middle of a coarse mesh time step, this is essential.